+++ backend/src/controllers/auth.controllers.js
@@ -1,5 +1,5 @@
 import User from '../models/user.models.js';
-import generateToken from '../helpers/generateToken.js';
+import generateCookie from '../helpers/generateCookie.js';
 
 export async function login(req, res) {
     try {
@@ -22,13 +22,10 @@ export async function login(req, res) {
             return res.status(403).json({ message: 'Invalid permissions' });
         }
 
-        res.cookie('token', generateToken(user._id), {
-            httpOnly: true,
-            secure: false,
-            sameSite: 'strict',
-        });
+        generateCookie(res, user._id);
 
-        res.json({ message: 'Logged' });
+        const { id, role } = user;
+        res.json({ id, username, role });
     } catch (err) {
         console.error(err);
         res.status(500).json({ message: err.message || 'Internal server error' });
@@ -67,5 +64,29 @@ export async function logout(req, res) {
             default:
                 return res.status(500).json({ message: err.message || 'Internal server error' });
         }
+    }
+}
+
+export async function refresh(req, res) {
+    try {
+        generateCookie(res, req.user._id);
+        res.status(200).json({ message: 'OK' });
+    } catch (err) {
+        console.error(err);
+        res.status(500).json({ message: err.message || 'Internal server error' });
+    }
+}
+
+export async function getMe(req, res) {
+    try {
+        const user = req.user;
+        if (!user) {
+            return res.json(null);
+        }
+        const { id, username, role } = user;
+        res.json({ id, username, role });
+    } catch (err) {
+        console.error(err);
+        res.status(500).json({ message: err.message || 'Internal server error' });
     }
 }