+++ backend/src/helpers/generateCookie.js
@@ -1,0 +1,9 @@
+import generateToken from './generateToken.js';
+
+export default function generateCookie(res, userId) {
+    res.cookie('token', generateToken(userId), {
+        httpOnly: true,
+        secure: false,
+        sameSite: 'strict',
+    });
+}