+++ frontend/src/types/accessorie.ts
@@ -1,0 +1,5 @@
+export default interface Accessorie {
+    id: string;
+    name: string;
+    quantity: number;
+}