+++ frontend/src/types/asset.ts
@@ -1,0 +1,4 @@
+export default interface Asset {
+    id: string;
+    name: string;
+}