+++ frontend/src/types/resource.ts
@@ -1,0 +1,5 @@
+import type ResourceResponse from './resourceResponse';
+
+export default interface Resource<T> {
+    get: () => Promise<ResourceResponse<T> | null>;
+}