+++ frontend/src/types/resourceResponse.ts
@@ -1,0 +1,4 @@
+export default interface ResourceResponse<T> {
+    headers: Array<String>;
+    data: Array<T>;
+}